requirements.txt
asgiref==3.11.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.57.0
django-cors-headers==4.3.0
django-extensions==4.1
django-filter==23.5
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
graphene==3.4.3
graphene-django==3.1.5
graphql-core==3.2.7
graphql-relay==3.2.0
idna==3.11
jwcrypto==1.5.6
mysqlclient==2.2.0
oauthlib==3.3.1
promise==2.3
pycparser==3.0
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.5.5
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
